Sample Data for Population C

The data contains survival information for boreal caribou population C which spans 11 years.

Usage

bbousurv_c

Format

A tibble with columns:

PopulationName

Name of the population

Year

The year the observation occurred

Month

The month the observation occurred

StartTotal

The total number of collared caribou at the start of the month

MortalitiesCertain

The number of confirmed caribou mortalities in the month

MortalitiesUncertain

The total number of mortalities that were not confirmed in that month

Details

The data is released under the Open Government Licence - Alberta